Moved parks into proper parklist
