+++ scripts/main.js
@@ -1,15 +1,5 @@
 console.log("Welcome to the main module")
 
-import { getParks, useParks } from "./parks/ParkProvider.js"
+import { ParksList } from "./parks/ParksProvider.js"
 
-const ParkList = () => {
-    getParks()
-        .then(() => {
-            const parks = useParks()
-            parks.forEach(element => {
-                console.log(parks.length +" "+ element.name +" "+ element.id)
-            });
-        })
-}
-
-ParkList()
+ParksList()
